+++ src/game/data/DocumentLoader.ts
@@ -1,3 +1,5 @@
+import { PlayerSaveManager } from '../systems/PlayerSaveData';
+
 export interface LibraryDocumentMetadata {
     id: string;
     title: string;
@@ -61,11 +63,14 @@ export async function loadAllDocuments(): Promise<void> {
 }
 
 /**
- * 全ての文書データを取得
+ * 全ての文書データを取得（最新の既読状態を反映）
  * @returns 文書データの配列
  */
 export function getAllDocuments(): LibraryDocument[] {
-    return Array.from(documentCache.values());
+    return Array.from(documentCache.values()).map(doc => ({
+        ...doc,
+        isRead: PlayerSaveManager.isDocumentRead(doc.id)
+    }));
 }
 
 /**
@@ -76,7 +81,11 @@ export function getAllDocuments(): LibraryDocument[] {
  */
 export function getDocument(id: string): LibraryDocument {
     if (documentCache.has(id)) {
-        return documentCache.get(id)!;
+        const doc = documentCache.get(id)!;
+        return {
+            ...doc,
+            isRead: PlayerSaveManager.isDocumentRead(doc.id)
+        };
     }
     throw new Error(`Document ${id} not loaded. Call loadAllDocuments() first.`);
 }
@@ -99,4 +108,40 @@ export function getUnreadDocumentIds(): string[] {
     return allDocuments
         .filter(doc => doc.unlocked && !doc.isRead)
         .map(doc => doc.id);
+}
+
+/**
+ * プレイヤーの状態に基づいて未読文書数を計算
+ * @param explorerLevel エクスプローラーレベル
+ * @param defeatedBosses 撃破済みボスID配列
+ * @param lostToBosses 敗北済みボスID配列
+ * @returns 未読文書数
+ */
+export function getUnreadCountForPlayer(
+    explorerLevel: number,
+    defeatedBosses: string[],
+    lostToBosses: string[]
+): number {
+    const allDocuments = getAllDocuments();
+    
+    // 解禁済み文書をフィルタ
+    const unlockedDocuments = allDocuments.filter(doc => {
+        const levelOk = !doc.requiredExplorerLevel || explorerLevel >= doc.requiredExplorerLevel;
+        let bossDefeatsOk = true;
+        if (doc.requiredBossDefeats && doc.requiredBossDefeats.length > 0) {
+            bossDefeatsOk = doc.requiredBossDefeats.every(bossId => 
+                defeatedBosses.includes(bossId)
+            );
+        }
+        let bossLossesOk = true;
+        if (doc.requiredBossLosses && doc.requiredBossLosses.length > 0) {
+            bossLossesOk = doc.requiredBossLosses.every(bossId => 
+                lostToBosses.includes(bossId)
+            );
+        }
+        return levelOk && bossDefeatsOk && bossLossesOk;
+    });
+
+    // 未読文書数を計算
+    return unlockedDocuments.filter(doc => !doc.isRead).length;
 }
